[fontir] Move StaticMetadata (& related) to module
No functional change, just cleaning this up a bit before I add stuff for
rvrn/FeatureVariations
